source/requirements.txt
nltk
tqdm
numpy
jinja2
pyyaml
openai>=1.0
tenacity
